move EqualsAvoidsNullVisitor into Recipe
